// frontend/src/api/generated/model/roadmapFeatureStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */

/**
 * Status of a roadmap feature
 */
export type RoadmapFeatureStatus = typeof RoadmapFeatureStatus[keyof typeof RoadmapFeatureStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RoadmapFeatureStatus = {
  under_review: 'under_review',
  planned: 'planned',
  in_progress: 'in_progress',
  done: 'done',
} as const;